codigo modularizado
